(feat) Make createEntryUrlBuilder() Craft 5 compatible
This PR bring some changes around the fact that Craft 5 now returns
complete, full url for entries. Since they are based on the cms, they
may not be want we want to use: We do not want the CMS to have the final
words on that.

Old api were deprecated.