(chore) Refactor the code around parsed and entry (#52)
* (chore) Refactor the code around parsed and entry

This commit changes a couple of things to make both apis be more
similar.

There was an error in parsed where toSchemeLess would return a
protocol-less version. This is fixed.

toRelative does not make sense without a 'base' url (relative to what?).
node's path.relative requires 2 params.
So it is now deprecated

4 new functions, across 3 new files were created to avoid code dup.
2 of them are exposed in the barrel file, since they work on strings.
The other two are not since they require a valid URL object, which is
not validated, hence, not exposed.

Some improvement, IMHO, to the docs.

* (chore) Add cs

* Fix typo

Co-authored-by: Felix Guerin <guerinfelix08@gmail.com>

* Fix typo

Co-authored-by: Felix Guerin <guerinfelix08@gmail.com>

---------

Co-authored-by: Felix Guerin <guerinfelix08@gmail.com>