(feat) Add fetchTimeout()
The default fetch api does not allow for a custom timeout value.
The default would then be left for the server to decide,
which is often 60 seconds. That way too long.

This change add a new wrapper with a 20 seconds default.

Since we have 60 seconds ourselve to generate a response (vercel's limit)
and we often have more than one fetch to issue, 20 seconds seems like a good default.