(feat) Add new site-router hook
This hook should be ran first, and is responsible to split the request's
url into more usable parts, which are the site uri and entry uri.

It centralize code we have been splitting across the codebases which
made it hard to customize this.

This implementation is the behavior we've been usign, which is to take
the first part of the pathname to be the site uri.

The main parts of the app that needs to be tied with this new hook are
i18n and graphql queries, to make sure they use the same information.

It also takes into account the fact that craft those not allow dashes in
site handles.