(feat) site uri vs i18n (#75)
* (feat) Add new site-router hook

This hook should be ran first, and is responsible to split the request's
url into more usable parts, which are the site uri and entry uri.

It centralize code we have been splitting across the codebases which
made it hard to customize this.

This implementation is the behavior we've been usign, which is to take
the first part of the pathname to be the site uri.

The main parts of the app that needs to be tied with this new hook are
i18n and graphql queries, to make sure they use the same information.

It also takes into account the fact that craft those not allow dashes in
site handles.

* (chore) Add hooks changeset

* (feat) Add acceptedLanguage

This is a refactor to get code out from getLangFromRequest() which is
now deprecated in favor of the new siteRouter hook.

* (chore) Add http changeset

* (feat) Use siteRouter values in i18n hooks

This will make sure that we are usign the proper values to find
locale/language information, based on the logic executed by the
siteRouteur hook.

The idea was to keep the behavior the same, but no longer use the
deprecated getLangFromRequest() function.

Bonus: Locals are now properly typed.

* (chore) Add i18n changeset

* (chore) Update lockfile

* (chore) Format code

* (fix) Import from js file

* (fix) event.url is already a URL

* (fix) Only set handle if falsy

* (fix) Type issue

* (chore) Add docs about the site router dependancy

* (chore) Format code

* (fix) Typing issue

* (chore) Format code

---------

Co-authored-by: nitriques <771169+nitriques@users.noreply.github.com>